regression-test :: flatten RegressionTest and RegressionTestUtil
